requirements.txt
pillow
requests
pyinstaller
yt_dlp